[shell-release]next version 3.1.7-SNAPSHOT

+++ admin-web/src/main/java/com/zrlog/admin/web/controller/api/AdminArticleController.java
@@ -3,9 +3,7 @@ package com.zrlog.admin.web.controller.api;
 import com.hibegin.common.util.BeanUtil;
 import com.hibegin.common.util.StringUtils;
 import com.hibegin.http.annotation.ResponseBody;
-import com.hibegin.http.server.web.Controller;
-import com.zrlog.admin.business.exception.ArgsException;
-import com.zrlog.admin.business.exception.NotFindDbEntryException;
+import com.zrlog.admin.business.AdminConstants;
 import com.zrlog.admin.business.exception.PermissionErrorException;
 import com.zrlog.admin.business.rest.request.CreateArticleRequest;
 import com.zrlog.admin.business.rest.request.UpdateArticleRequest;
@@ -13,41 +11,45 @@ import com.zrlog.admin.business.rest.response.*;
 import com.zrlog.admin.business.service.AdminArticleService;
 import com.zrlog.admin.web.annotation.RefreshCache;
 import com.zrlog.admin.web.token.AdminTokenThreadLocal;
-import com.zrlog.blog.web.util.ControllerUtil;
-import com.zrlog.blog.web.util.WebTools;
-import com.zrlog.business.rest.response.ArticleResponseEntry;
-import com.zrlog.business.service.TemplateHelper;
+import com.zrlog.business.plugin.StaticSitePlugin;
+import com.zrlog.business.service.VisitorArticleService;
+import com.zrlog.business.util.ControllerUtil;
 import com.zrlog.common.Constants;
-import com.zrlog.common.ZrLogConfig;
-import com.zrlog.common.rest.response.ApiStandardResponse;
-import com.zrlog.data.dto.PageData;
+import com.zrlog.common.controller.BaseController;
+import com.zrlog.common.exception.ArgsException;
+import com.zrlog.common.exception.NotFindDbEntryException;
 import com.zrlog.model.Log;
-import com.zrlog.model.Tag;
-import com.zrlog.model.Type;
 import com.zrlog.util.I18nUtil;
+import com.zrlog.util.ThreadUtils;
 import com.zrlog.util.ZrLogUtil;
 
 import java.sql.SQLException;
 import java.util.Map;
 import java.util.Objects;
 import java.util.StringJoiner;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
 
-public class AdminArticleController extends Controller {
+public class AdminArticleController extends BaseController {
 
     private final AdminArticleService articleService = new AdminArticleService();
 
     @RefreshCache(async = true)
     @ResponseBody
-    public ApiStandardResponse<DeleteLogResponse> delete() throws SQLException {
+    public AdminApiPageDataStandardResponse<DeleteLogResponse> delete() throws SQLException {
         if (ZrLogUtil.isPreviewMode()) {
             throw new PermissionErrorException();
         }
-        String[] ids = getRequest().getParaToStr("id").split(",");
+        String idStr = getParamWithEmptyCheck("id");
+        if (StringUtils.isEmpty(idStr)) {
+            throw new ArgsException("id");
+        }
+        String[] ids = idStr.split(",");
         for (String id : ids) {
             articleService.delete(Long.valueOf(id));
         }
-        return new ApiStandardResponse<>(new DeleteLogResponse(true));
+        return new AdminApiPageDataStandardResponse<>(new DeleteLogResponse(true));
     }
 
     private String getResponseMsg(CreateOrUpdateArticleResponse response) {
@@ -55,72 +57,84 @@ public class AdminArticleController extends Controller {
                 || Objects.equals(response.getPrivacy(), true) ? "saveSuccess" : "releaseSuccess");
     }
 
-    @RefreshCache(async = true)
     @ResponseBody
-    public ApiStandardResponse<CreateOrUpdateArticleResponse> create() {
-        CreateOrUpdateArticleResponse create = articleService.create(AdminTokenThreadLocal.getUser(), BeanUtil.convertWithValid(getRequest().getInputStream(),
-                CreateArticleRequest.class));
-        create.setPreviewUrl(getPreviewUrl(create));
-        return new ApiStandardResponse<>(create, getResponseMsg(create));
+    public AdminApiPageDataStandardResponse<ArticleGlobalResponse> create() throws SQLException {
+        CreateOrUpdateArticleResponse create = articleService.create(AdminTokenThreadLocal.getUser(),
+                getRequestBodyWithNullCheck(CreateArticleRequest.class));
+        AdminApiPageDataStandardResponse<ArticleGlobalResponse> apiResponse = loadDetailById(create.getLogId() + "");
+        apiResponse.setMessage(getResponseMsg(create));
+        return apiResponse;
     }
 
     @ResponseBody
-    public ApiStandardResponse<CreateOrUpdateArticleResponse> update() {
-        CreateOrUpdateArticleResponse update = articleService.update(AdminTokenThreadLocal.getUser(), BeanUtil.convertWithValid(getRequest().getInputStream(),
-                UpdateArticleRequest.class));
-        update.setPreviewUrl(getPreviewUrl(update));
+    public AdminApiPageDataStandardResponse<ArticleGlobalResponse> update() throws SQLException {
+        CreateOrUpdateArticleResponse update = articleService.update(AdminTokenThreadLocal.getUser(),
+                getRequestBodyWithNullCheck(UpdateArticleRequest.class));
+        AdminApiPageDataStandardResponse<ArticleGlobalResponse> apiResponse = loadDetailById(update.getLogId() + "");
+        LoadEditArticleResponse loadEditArticleResponse = apiResponse.getData().getArticle();
         //为发布状态才需要更新缓存信息（避免无用更新）
-        if (Objects.nonNull(update.getRubbish()) && Objects.equals(update.getRubbish(), false)) {
-            Constants.zrLogConfig.getCacheService().refreshInitDataCacheAsync(request, true);
+        if (Objects.equals(loadEditArticleResponse.isRubbish(), false)) {
+            request.getAttr().put(AdminConstants.SYNC_UPDATE_CACHE_KEY, true);
         }
-        return new ApiStandardResponse<>(update, getResponseMsg(update));
+        apiResponse.setMessage(getResponseMsg(update));
+        return apiResponse;
     }
 
     @ResponseBody
     public AdminApiPageDataStandardResponse<ArticlePageData> index() throws SQLException, ExecutionException, InterruptedException {
-        String key = WebTools.convertRequestParam(request.getParaToStr("key"));
-        String types = WebTools.convertRequestParam(request.getParaToStr("types"));
-        int articlePageSize = Constants.getAdminArticlePageSize();
+        String key = request.getParaToStr("key", "");
+        String types = request.getParaToStr("types", "");
+        int articlePageSize = AdminConstants.getAdminArticlePageSize();
         ArticlePageData pageData = articleService.adminPage(ControllerUtil.toPageRequest(this, articlePageSize), key, types, request);
-        pageData.setKey(key);
-        pageData.setDefaultPageSize(Long.parseLong(articlePageSize + ""));
-        pageData.setTypes(Constants.zrLogConfig.getCacheService().refreshInitDataCacheAsync(request, false).get().getTypes());
         AdminApiPageDataStandardResponse<ArticlePageData> standardResponse = new AdminApiPageDataStandardResponse<>(pageData);
-        standardResponse.setDocumentTitle(Constants.getAdminTitle(I18nUtil.getAdminStringFromRes("blogManage")));
+        standardResponse.setDocumentTitle(AdminConstants.getAdminDocumentTitleByUri(request.getUri()));
         return standardResponse;
     }
 
 
-    private String getPreviewUrl(CreateOrUpdateArticleResponse articleResponseEntry) {
+    private String getPreviewUrl(LoadEditArticleResponse articleResponseEntry) {
         String key = articleResponseEntry.getLogId() + "";
         if (StringUtils.isNotEmpty(articleResponseEntry.getAlias())) {
             key = articleResponseEntry.getAlias();
         }
-        if (articleResponseEntry.getPrivacy() || articleResponseEntry.getRubbish()) {
+        if (articleResponseEntry.isPrivacy() || articleResponseEntry.isRubbish()) {
             return "/admin/403?message=" + I18nUtil.getBackendStringFromRes("preview403");
         }
-        return ZrLogUtil.getHomeUrlWithHost(getRequest()) + Constants.getArticleUri() + key + TemplateHelper.getSuffix(request);
+        return ZrLogUtil.getHomeUrlWithHost(getRequest()) + Constants.getArticleUri() + key + StaticSitePlugin.getSuffix(request);
     }
 
-    @ResponseBody
-    public AdminApiPageDataStandardResponse<ArticleGlobalResponse> articleEdit() throws SQLException {
+    private AdminApiPageDataStandardResponse<ArticleGlobalResponse> loadDetailById(String id) throws SQLException {
         ArticleGlobalResponse response = new ArticleGlobalResponse();
-        response.setTags(new Tag().findAll());
-        response.setTypes(new Type().findAll());
-        String id = request.getParaToStr("id");
+        ExecutorService executorService = ThreadUtils.newFixedThreadPool(2);
+        try {
+            CompletableFuture.allOf(CompletableFuture.runAsync(() -> {
+                response.setTags(Constants.zrLogConfig.getCacheService().getTags());
+            }, executorService), CompletableFuture.runAsync(() -> {
+                response.setTypes(Constants.zrLogConfig.getCacheService().getArticleTypes());
+
+            }, executorService)).join();
+        } finally {
+            executorService.shutdown();
+        }
         if (StringUtils.isNotEmpty(id)) {
-            response.setArticle(detail().getData());
+            response.setArticle(loadDetail(id));
         } else {
             response.setArticle(new LoadEditArticleResponse());
         }
         AdminApiPageDataStandardResponse<ArticleGlobalResponse> standardResponse = new AdminApiPageDataStandardResponse<>(response);
-        StringJoiner sj = new StringJoiner(Constants.ADMIN_TITLE_CHAR);
+        StringJoiner sj = new StringJoiner(AdminConstants.ADMIN_TITLE_CHAR);
         if (Objects.nonNull(response.getArticle().getTitle())) {
             sj.add(response.getArticle().getTitle());
         }
-        sj.add(I18nUtil.getAdminStringFromRes("admin.log.edit"));
-        standardResponse.setDocumentTitle(Constants.getAdminTitle(sj.toString()));
+        sj.add(AdminConstants.getAdminDocumentTitleByUri(request.getUri()));
+        standardResponse.setDocumentTitle(sj.toString());
         return standardResponse;
+    }
+
+    @ResponseBody
+    public AdminApiPageDataStandardResponse<ArticleGlobalResponse> articleEdit() throws SQLException {
+        String id = request.getParaToStr("id", "");
+        return loadDetailById(id);
     }
 
     /**
@@ -130,11 +144,12 @@ public class AdminArticleController extends Controller {
      * @throws SQLException
      */
     @ResponseBody
-    public ApiStandardResponse<LoadEditArticleResponse> detail() throws SQLException {
-        String id = getRequest().getParaToStr("id");
-        if (StringUtils.isEmpty(id)) {
-            throw new ArgsException("id");
-        }
+    @Deprecated
+    public AdminApiPageDataStandardResponse<LoadEditArticleResponse> detail() throws SQLException {
+        return new AdminApiPageDataStandardResponse<>(loadDetail(getParamWithEmptyCheck("id")));
+    }
+
+    private LoadEditArticleResponse loadDetail(String id) throws SQLException {
         Map<String, Object> log = new Log().adminFindByIdOrAlias(id);
         if (log == null) {
             throw new NotFindDbEntryException();
@@ -142,7 +157,9 @@ public class AdminArticleController extends Controller {
         log.remove("releaseTime");
         log.remove("last_update_date");
         log.remove("lastUpdateDate");
-        return new ApiStandardResponse<>(BeanUtil.convert(log, LoadEditArticleResponse.class));
+        LoadEditArticleResponse loadEditArticleResponse = BeanUtil.convert(VisitorArticleService.handlerArticle(log), LoadEditArticleResponse.class);
+        loadEditArticleResponse.setPreviewUrl(getPreviewUrl(loadEditArticleResponse));
+        return loadEditArticleResponse;
     }
 
 }
